requirements.txt
selenium == 4.29.0
python-dotenv
anthropic
google-auth
requests
webdriver-manager == 4.0.2
psutil
Pillow
google-cloud-videointelligence
openai